requirements.txt
langchain
langchain_google_genai
requests
flask
flask-cors
python-dotenv
pandas